+++ backend/src/baserow/contrib/database/api/urls.py
@@ -8,6 +8,8 @@ from .rows import urls as row_urls
 from .tokens import urls as token_urls
 from .export import urls as export_urls
 from .formula import urls as formula_urls
+from .airtable import urls as airtable_urls
+
 
 app_name = "baserow.contrib.database.api"
 
@@ -20,4 +22,5 @@ urlpatterns = [
     path("tokens/", include(token_urls, namespace="tokens")),
     path("export/", include(export_urls, namespace="export")),
     path("formula/", include(formula_urls, namespace="formula")),
+    path("airtable/", include(airtable_urls, namespace="airtable")),
 ]