+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Added accept `image/*` attribute to the form cover and logo upload. 
 * Added management to import a shared Airtable base.
+* Added web-frontend interface to import a shared Airtable base.
 * Fixed adding new fields in the edit row popup that require refresh in Kanban and Form views.
 * Cache model fields when generating model.
 * Fixed `'<' not supported between instances of 'NoneType' and 'int'` error. Blank